Filters - QA. for now - server don't supposed to crush

+++ BackEnd_Files/api/controllers/filters.js
@@ -141,8 +141,7 @@ function interestedInHelper_AccordingToGenderAndSexualOrientation(
 			sqlQuery = `(sexual_orientation like 'Asexual')`;
 			break;
 		default:
-			sqlQuery =
-				"ILLEGAL function 'interestedInHelper_AccordingToGenderAndSexualOrientation' - switch case --> default";
+			sqlQuery = undefined;
 	}
 
 	return cb(sqlQuery);
@@ -175,6 +174,9 @@ function interestedInHelper_AccordingToUserInterestedInChosen(
 			currentUserGender,
 			currentUserSexualOrientation,
 			(toConcatSqlQuery) => {
+				if (toConcatSqlQuery === undefined) {
+					return cb(undefined);
+				}
 				sqlQuery = " and (";
 				sqlQuery = sqlQuery.concat(toConcatSqlQuery);
 				sqlQuery = sqlQuery.concat(
@@ -248,7 +250,7 @@ function createInterestedInQuery(req) {
 							currentUserGender,
 							currentUserSexualOrientation,
 							(sqlQuery) => {
-								if (sqlQuery.length <= 0) {
+								if (sqlQuery === undefined || sqlQuery.length <= 0) {
 									reject(
 										`!!INTERESTED IN!!\nSomething went wrong. Interested In parameter illegal`
 									);
@@ -369,9 +371,16 @@ getAllFilters = (req, res) => {
 	logger.info("This is an info log");
 	mySqlConnection.query("SELECT * FROM Filters", (err, rows) => {
 		try {
-			res.send(rows);
+			if (err || rows === undefined) {
+				throw new Error(
+					"Filters - GET request, get all filters in DB. MySQL Error"
+				);
+			} else {
+				res.send(rows);
+			}
 		} catch (err) {
-			console.log(err.message);
+			logger.error({err});
+			return res.status(500).send("internal error");
 		}
 	});
 };
@@ -384,11 +393,7 @@ getUserFilter = (req, cb) => {
 		"SELECT * FROM Filters WHERE user_id = ?",
 		[userid],
 		(err, rows) => {
-			try {
-				return cb(rows);
-			} catch (err) {
-				console.log(err.message);
-			}
+			return cb(rows);
 		}
 	);
 };
@@ -405,84 +410,123 @@ getFriendsOfFriends = (req, res) => {
 		`select user_A_id, user_B_id from Connections where (user_A_id = ${userid} or user_B_id = ${userid}) and connected = 1`,
 		(err, rows) => {
 			try {
-				if (rows !== undefined && rows.length > 0) {
-					for (row = 0; row < rows.length; row++) {
-						if (parseInt(rows[row].user_a_id, 10) !== parseInt(userid, 10)) {
-							myFriendsUserid.push(parseInt(rows[row].user_a_id, 10));
-						} else {
-							myFriendsUserid.push(parseInt(rows[row].user_b_id, 10));
+				if (err || rows === undefined) {
+					throw new Error(
+						"Filters - friends of friends. GET request, MySQL Error"
+					);
+				} else {
+					if (rows.length > 0) {
+						for (row = 0; row < rows.length; row++) {
+							if (parseInt(rows[row].user_A_id, 10) !== parseInt(userid, 10)) {
+								myFriendsUserid.push(parseInt(rows[row].user_A_id, 10));
+							} else {
+								myFriendsUserid.push(parseInt(rows[row].user_B_id, 10));
+							}
 						}
-					}
-					createFriendsOfFriendsQuery(myFriendsUserid, sqlQuery, (response) => {
-						mySqlConnection.query(response, (err, rows) => {
-							try {
-								if (rows !== undefined && rows.length > 0) {
-									for (
-										myFriendOfFriend = 0;
-										myFriendOfFriend < rows.length;
-										myFriendOfFriend++
-									) {
-										if (
-											!myFriendsUserid.includes(
-												parseInt(rows[myFriendOfFriend].user_A_id, 10)
-											)
-										) {
-											myFriendsOfFriendsUserid.push(
-												parseInt(rows[myFriendOfFriend].user_A_id, 10)
-											);
-										}
-										if (
-											!myFriendsUserid.includes(
-												parseInt(rows[myFriendOfFriend].user_B_id, 10)
-											)
-										) {
-											myFriendsOfFriendsUserid.push(
-												parseInt(rows[myFriendOfFriend].user_B_id, 10)
-											);
-										}
-									}
+
+						createFriendsOfFriendsQuery(
+							myFriendsUserid,
+							sqlQuery,
+							(response) => {
+								if (response === undefined) {
+									throw new Error(
+										"Filters - friends of friends. Create friends of friends function cb failed"
+									);
 								}
 
-								let myFriendsRequest = {
-									params: {
-										userid: String(myFriendsUserid),
-									},
-								};
-								let myFriendsOfFriendsRequest = {
-									params: {
-										userid: String(myFriendsOfFriendsUserid),
-									},
-								};
-								getUserConfigurationInner(
-									myFriendsRequest,
-									userid,
-									(myFriendsConfigurations) => {
-										Object.assign(myFriends_myFriendsOfFriends, {
-											myFriends: myFriendsConfigurations,
-										});
-
-										getUserConfigurationInner(
-											myFriendsOfFriendsRequest,
-											userid,
-											(myFriendsOfFriendsConfigurations) => {
-												Object.assign(myFriends_myFriendsOfFriends, {
-													myFriendsOfFriends: myFriendsOfFriendsConfigurations,
-												});
-												res.send(myFriends_myFriendsOfFriends);
+								console.log(response);
+								mySqlConnection.query(response, (err, rows) => {
+									try {
+										if (err || rows === undefined) {
+											throw new Error(
+												"Filters - friends of friends. MySQL Error"
+											);
+										} else {
+											if (rows.length > 0) {
+												for (
+													myFriendOfFriend = 0;
+													myFriendOfFriend < rows.length;
+													myFriendOfFriend++
+												) {
+													if (
+														!myFriendsUserid.includes(
+															parseInt(rows[myFriendOfFriend].user_A_id, 10)
+														)
+													) {
+														myFriendsOfFriendsUserid.push(
+															parseInt(rows[myFriendOfFriend].user_A_id, 10)
+														);
+													}
+													if (
+														!myFriendsUserid.includes(
+															parseInt(rows[myFriendOfFriend].user_B_id, 10)
+														)
+													) {
+														myFriendsOfFriendsUserid.push(
+															parseInt(rows[myFriendOfFriend].user_B_id, 10)
+														);
+													}
+												}
 											}
-										);
+
+											let myFriendsRequest = {
+												params: {
+													userid: String(myFriendsUserid),
+												},
+											};
+											let myFriendsOfFriendsRequest = {
+												params: {
+													userid: String(myFriendsOfFriendsUserid),
+												},
+											};
+											getUserConfigurationInner(
+												myFriendsRequest,
+												userid,
+												(myFriendsConfigurations) => {
+													if (myFriendsConfigurations === undefined) {
+														throw new Error(
+															"Filters - friends of friends. **MY FRIENDS CONFIGURATIONS**\nGet Users Configuration Inner (cb) failed"
+														);
+													}
+
+													Object.assign(myFriends_myFriendsOfFriends, {
+														myFriends: myFriendsConfigurations,
+													});
+
+													getUserConfigurationInner(
+														myFriendsOfFriendsRequest,
+														userid,
+														(myFriendsOfFriendsConfigurations) => {
+															if (myFriendsConfigurations === undefined) {
+																throw new Error(
+																	"Filters - friends of friends. **MY FRIENDS OF FRIENDS CONFIGURATIONS**\nGet Users Configuration Inner (cb) failed"
+																);
+															}
+
+															Object.assign(myFriends_myFriendsOfFriends, {
+																myFriendsOfFriends:
+																	myFriendsOfFriendsConfigurations,
+															});
+															res.send(myFriends_myFriendsOfFriends);
+														}
+													);
+												}
+											);
+										}
+									} catch (err) {
+										logger.error({err});
+										return res.status(500).send("internal error");
 									}
-								);
-							} catch (err) {
-								console.log(err.message);
+								});
 							}
-						});
-					});
-				} else {
-					return res.send(rows);
+						);
+					} else {
+						return res.send(rows);
+					}
 				}
 			} catch (err) {
-				console.log(err.message);
+				logger.error({err});
+				return res.status(500).send("internal error");
 			}
 		}
 	);
@@ -499,6 +543,11 @@ getUserFilteredUsers = (req, res) => {
 	let resolve;
 
 	getUserFilter(req, async (userFilter) => {
+		if (userFilter === undefined) {
+			logger.error("Filters - GET a user filters, MySQL Error");
+			return res.status(500).send(`Internal Error`);
+		}
+
 		if (userFilter.length === 0) {
 			// user don't have filters - return to client all users in db
 			getUserFilteredUsers_OnlyOnline_Helper(
@@ -624,16 +673,20 @@ deleteUserFilter = (req, res) => {
 	const userid = req.params.userid;
 
 	mySqlConnection.query(
-		"delete from Filters where user_id = ?",
-		[userid],
-		(err, result) => {
+		`delete from Filters where user_id = ${userid}`,
+		(err, rows) => {
 			try {
-				msgToClient = {
-					msg: `User number ${userid} filter deleted successfully`,
-				};
-				return res.send(msgToClient);
+				if (err || rows === undefined) {
+					throw new Error("Filters - DELETE request. MySQL Error");
+				} else {
+					msgToClient = {
+						msg: `User number ${userid} filter deleted successfully`,
+					};
+					return res.send(msgToClient);
+				}
 			} catch (err) {
-				console.log(err.message);
+				logger.error({err});
+				return res.status(500).send(`Internal Error\n${err}`);
 			}
 		}
 	);
